Added comments; placed 'onBackButtonTapped' code into LayoutManager class (all layouts now inherit the functionality)
