+++ main-library/src/main/java/rocket/agile/com/mainlibrary/activity/LayoutView_Buttons_Grid.java
@@ -3,11 +3,17 @@ package rocket.agile.com.mainlibrary.activity;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
-
 import rocket.agile.com.mainlibrary.R;
+
+/**
+ * Created by keithkowalski on 6/19/17.
+ *
+ * Purpose:  Present grid layout view to users
+ * Function: A) Grid of buttons is created, based on action items set by admin
+ *
+ */
 
 public class LayoutView_Buttons_Grid extends LayoutManager {
